init Location model
