+++ packages/database/src/models/Facility.ts
@@ -1,0 +1,3 @@
+import { Model } from './Model';
+
+export class Facility extends Model {}
