+++ packages/database/src/models/ImagingRequest.ts
@@ -1,0 +1,3 @@
+import { Model } from './Model';
+
+export class ImagingRequest extends Model {}
