+++ packages/database/src/models/Procedure.ts
@@ -1,0 +1,3 @@
+import { Model } from './Model';
+
+export class Procedure extends Model {}
